certification

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(15);
+                    
         return view('tasks.index',[
             'tasks' => $tasks
         ]);
+        }
+        
+        return view('tasks.index',[
+            'tasks' => []
+        ]);
+        
     }
 
     /**
@@ -29,9 +38,13 @@ class TasksController extends Controller
     {
         $task = new Task;
         
-        return view('tasks.create',[
-            'task' => $task,
-        ]);
+        if (\Auth::check()){
+            return view('tasks.create',[
+                'task' => $task,
+            ]);
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -46,11 +59,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required',
         ]);
-        $task = new Task;
         
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
     }
@@ -65,9 +78,13 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        return view('tasks.show',[
-                'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.show',[
+                    'task' => $task,
+            ]);
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -80,9 +97,13 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        return view('tasks.edit', [
-            'task' => $task,    
-        ]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.edit', [
+                'task' => $task,    
+            ]);
+        }
+        
+        return redirect('/');
     }
 
     /**
@@ -94,15 +115,21 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'status' => 'required|max:10',
-            'content' => 'required',
-        ]);
         $task = Task::findOrFail($id);
         
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        if (\Auth::id() === $task->user_id) {
+            
+            $request->validate([
+                'status' => 'required|max:10',
+                'content' => 'required',
+            ]);
+            
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            
+            return redirect('/');
+        }
         
         return redirect('/');
     }
@@ -116,8 +143,14 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) {
         $task->delete();
         
-        return redirect('/');
+        return redirect('/')->with('success','Delete Successful');
+        }
+        
+        return redirect('/')
+            ->with('Delete Failed');
     }
 }
